notes-app-backend/requirements.txt
flask
flask-restful
flask-cors
pymongo
python-dotenv
bcrypt
pyjwt
flask-pymongo